Backend/requirements.txt
flask
flask-cors
torch
torchvision
faiss-cpu
Pillow
requests
tqdm
numpy
gunicorn